+++ Core/src/main/java/tify/server/core/consts/Status.java
@@ -1,0 +1,14 @@
+package tify.server.core.consts;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Status {
+    Y("true"),
+    N("false");
+
+    final String value;
+}
