+++ Domain/src/main/java/tify/server/domain/domains/alarm/adaptor/AlarmHistoryAdaptor.java
@@ -1,10 +1,11 @@
-package tify.server.domain.domains.user.adaptor;
+package tify.server.domain.domains.alarm.adaptor;
+
 
 import lombok.RequiredArgsConstructor;
 import tify.server.core.annotation.Adaptor;
-import tify.server.domain.domains.user.domain.AlarmHistory;
-import tify.server.domain.domains.user.exception.AlarmHistoryNotFoundException;
-import tify.server.domain.domains.user.repository.AlarmHistoryRepository;
+import tify.server.domain.domains.alarm.domain.AlarmHistory;
+import tify.server.domain.domains.alarm.exception.AlarmHistoryNotFoundException;
+import tify.server.domain.domains.alarm.repository.AlarmHistoryRepository;
 
 @Adaptor
 @RequiredArgsConstructor
@@ -18,5 +19,6 @@ public class AlarmHistoryAdaptor {
     }
 
     public AlarmHistory save(AlarmHistory alarmHistory) {
-        return alarmHistoryRepository.save(alarmHistory);}
+        return alarmHistoryRepository.save(alarmHistory);
+    }
 }
