+++ Domain/src/main/java/tify/server/domain/domains/alarm/domain/AlarmType.java
@@ -1,4 +1,5 @@
-package tify.server.domain.domains.user.domain;
+package tify.server.domain.domains.alarm.domain;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +8,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum AlarmType {
     ALARM("쿡 찌르기");
-//    ALARM("쿡 찌르기", false);
 
     final String type;
-//    final boolean checked;
 }
