+++ Domain/src/main/java/tify/server/domain/domains/alarm/exception/AlarmHistoryNotFoundException.java
@@ -1,4 +1,5 @@
-package tify.server.domain.domains.user.exception;
+package tify.server.domain.domains.alarm.exception;
+
 
 import tify.server.core.exception.BaseException;
 
@@ -6,5 +7,6 @@ public class AlarmHistoryNotFoundException extends BaseException {
     public static final BaseException EXCEPTION = new AlarmHistoryNotFoundException();
 
     private AlarmHistoryNotFoundException() {
-        super(AlarmHistoryException.ALARM_NOT_FOUND_ERROR);}
+        super(AlarmHistoryException.ALARM_NOT_FOUND_ERROR);
+    }
 }
