+++ Domain/src/main/java/tify/server/domain/domains/alarm/repository/AlarmHistoryRepository.java
@@ -1,6 +1,7 @@
-package tify.server.domain.domains.user.repository;
+package tify.server.domain.domains.alarm.repository;
+
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import tify.server.domain.domains.user.domain.AlarmHistory;
+import tify.server.domain.domains.alarm.domain.AlarmHistory;
 
 public interface AlarmHistoryRepository extends JpaRepository<AlarmHistory, Long> {}
