+++ Domain/src/main/java/tify/server/domain/domains/question/adaptor/AnswerAdaptor.java
@@ -1,11 +1,11 @@
-package tify.server.domain.domains.user.adaptor;
+package tify.server.domain.domains.question.adaptor;
 
 
 import lombok.RequiredArgsConstructor;
 import tify.server.core.annotation.Adaptor;
-import tify.server.domain.domains.user.domain.Answer;
-import tify.server.domain.domains.user.exception.AnswerNotFoundException;
-import tify.server.domain.domains.user.repository.AnswerRepository;
+import tify.server.domain.domains.question.domain.Answer;
+import tify.server.domain.domains.question.exception.AnswerNotFoundException;
+import tify.server.domain.domains.question.repository.AnswerRepository;
 
 @Adaptor
 @RequiredArgsConstructor
@@ -19,5 +19,7 @@ public class AnswerAdaptor {
                 .orElseThrow(() -> AnswerNotFoundException.EXCEPTION);
     }
 
-    public Answer save(Answer answer) { return answerRepository.save(answer); }
+    public Answer save(Answer answer) {
+        return answerRepository.save(answer);
+    }
 }
