+++ Domain/src/main/java/tify/server/domain/domains/question/domain/Answer.java
@@ -1,4 +1,4 @@
-package tify.server.domain.domains.user.domain;
+package tify.server.domain.domains.question.domain;
 
 
 import com.esotericsoftware.kryo.serializers.FieldSerializer.NotNull;
