+++ Domain/src/main/java/tify/server/domain/domains/question/domain/DailyQuestion.java
@@ -1,4 +1,4 @@
-package tify.server.domain.domains.user.domain;
+package tify.server.domain.domains.question.domain;
 
 
 import com.esotericsoftware.kryo.serializers.FieldSerializer.NotNull;
@@ -9,6 +9,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import tify.server.domain.domains.AbstractTimeStamp;
+import tify.server.domain.domains.user.domain.Category;
 
 @Getter
 @Entity
