+++ Domain/src/main/java/tify/server/domain/domains/question/domain/Knock.java
@@ -1,13 +1,13 @@
-package tify.server.domain.domains.user.domain;
+package tify.server.domain.domains.question.domain;
+
 
 import com.esotericsoftware.kryo.serializers.FieldSerializer.NotNull;
+import javax.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import tify.server.domain.domains.AbstractTimeStamp;
-
-import javax.persistence.*;
 
 @Getter
 @Entity
