+++ Domain/src/main/java/tify/server/domain/domains/question/exception/KnockNotFoundException.java
@@ -1,4 +1,5 @@
-package tify.server.domain.domains.user.exception;
+package tify.server.domain.domains.question.exception;
+
 
 import tify.server.core.exception.BaseException;
