+++ Domain/src/main/java/tify/server/domain/domains/question/repository/AnswerRepository.java
@@ -1,7 +1,7 @@
-package tify.server.domain.domains.user.repository;
+package tify.server.domain.domains.question.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import tify.server.domain.domains.user.domain.Answer;
+import tify.server.domain.domains.question.domain.Answer;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {}
