+++ Domain/src/main/java/tify/server/domain/domains/question/repository/DailyQuestionRepository.java
@@ -1,7 +1,7 @@
-package tify.server.domain.domains.user.repository;
+package tify.server.domain.domains.question.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import tify.server.domain.domains.user.domain.DailyQuestion;
+import tify.server.domain.domains.question.domain.DailyQuestion;
 
 public interface DailyQuestionRepository extends JpaRepository<DailyQuestion, Long> {}
