+++ Domain/src/main/java/tify/server/domain/domains/question/repository/KnockRepository.java
@@ -1,0 +1,7 @@
+package tify.server.domain.domains.question.repository;
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import tify.server.domain.domains.question.domain.Knock;
+
+public interface KnockRepository extends JpaRepository<Knock, Long> {}
