--- Domain/src/main/java/tify/server/domain/domains/user/repository/KnockRepository.java
@@ -1,7 +1,0 @@
-package tify.server.domain.domains.user.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import tify.server.domain.domains.user.domain.Knock;
-
-public interface KnockRepository extends JpaRepository<Knock, Long> {
-}
